+++ mybatis-plus-extension/src/main/java/com/baomidou/mybatisplus/extension/conditions/query/ChainQuery.java
@@ -63,7 +63,7 @@ public interface ChainQuery<T> extends ChainWrapper<T> {
      *
      * @return count
      */
-    default Integer count() {
+    default Long count() {
         return SqlHelper.retCount(getBaseMapper().selectCount(getWrapper()));
     }
